refactor gen state
